Error when rendering metadata editors fields with file attachments
This version of the code tries to add media uploads to all metadatafields in the metadataset editor view. 

The page breaks with a recursion error when adding a field to the metadata set.

When the Attachment variable has the [ScriptIgnore] attribute the page does finish rendering. When the [ScriptIgnore] attribute is removed but applied to inner variables for both Attachmnet and Formfield the page breaks.

The next step could be working on XmlModel but at that point something else may break.