Use FileDescription instead of DataFile
This is the first step towards using the FileDescription wrapper instead of using DataFile directly on the FormFieldViewModel. This allows to add elements to a file description like labels. Right now each file is saved and transformed to a FileDescription but the DataFile information is not being added to the XML.