Add automatically information about user creating an CFXmlModel
Change CFXmlModel so that when a new instance is created the attributes created-by-name and created-by-guid are set using the http context. This change had to be implemented in two parts; Catfish.core added the attributes and a static action to be set on Catfish. Catfish provides an action which sets the values that can only be retrieved using http context.