Fix BaseIntegrationTest and finish Entity creation test
The FindElement WebDriver extension would not always work. Added a
hack that will wait for 100 milliseconds before attempting to find
the requested element. The timeoutInSeconds wait should have taken
care of this but it still fails in some cases. It seems there is a
race condition happening.

This fix also completes the EntityType creation test.

We should take another look at the FindElement WebDriver extension
to remove this hack.