WIP matching solr information
At this stage we are starting to verify the information stored in solr matches the intended output.

Change in direction: A new requirement came up that specifies solr values for metadata_fields will need to be changed to values as list of strings _ss. This will simplify a lot of thigs.